Project completed works fine

+++ controllers/contactController.js
@@ -2,15 +2,15 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel");
 //@desc Get all contacts
 //@route Get  /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contacts);
 });
 
 //@desc Create New contacts
 //@route POST  /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
   console.log(req.body);
   const { name, email, phone } = req.body;
@@ -23,13 +23,14 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
   res.status(200).json(contact);
 });
 
 //@desc Get contacts
 //@route GET  /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
@@ -40,13 +41,18 @@ const getContact = asyncHandler(async (req, res) => {
 });
 
 //@desc Update contacts
-//@route GET  /api/contacts/:id
-//@access public
+//@route PUT  /api/contacts/:id
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("contact not found");
+  }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error(" User don't have permision to update other user contacts");
   }
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
@@ -57,12 +63,17 @@ const updateContact = asyncHandler(async (req, res) => {
 });
 //@desc Delete contacts
 //@route DELETE  /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("contact not found");
+  }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error(" User dont' have permision to update other user contacts");
   }
   await Contact.deleteOne();
   res.status(200).json(contact);
